homework 5
