+++ js/data.js
@@ -1,0 +1,66 @@
+const DESCRIPTION = [
+  'Город в котором мы побывали',
+  'Хуан Себастьян какой-то человек',
+  'Мария - красивая девушка',
+  'Большая река',
+  'Высокие горы',
+  'Теплая вода',
+  'Люпита',
+  'Вашингтон',
+  'да Марья',
+  'Верон',
+  'Мирабелла',
+  'Вальц',
+  'Онопко',
+  'Топольницкая',
+  'Нионго',
+  'Ирвинг',
+  'black',
+  'red',
+  'blue',
+  'yellow',
+  'green',
+  'rgb(146, 100, 161)',
+  'rgb(56, 159, 117)',
+  'rgb(215, 210, 55)',
+  'rgb(0, 0, 0)',
+];
+
+const userName = [
+  'Иван',
+  'Хуан Себастьян',
+  'Мария',
+  'Кристоф',
+  'Виктор',
+  'Юлия',
+  'Люпита',
+  'Вашингтон',
+  'да Марья',
+  'Верон',
+  'Мирабелла',
+  'Вальц',
+  'Онопко',
+  'Топольницкая',
+  'Нионго',
+  'Ирвинг',
+  'Дмитрий',
+  'Николай',
+  'Татьяна',
+  'Алена',
+  'Ольга',
+  'Марина',
+  'Стас',
+  'Роналдо'
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+
+export { DESCRIPTION, userName, MESSAGE };
